requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
requests==2.31.0
google-generativeai>=0.8.0
python-dotenv==1.0.0
